Skeleton model framework provided by OWL2MAS. Applications extend the AgentModel class with agent types specific to their application. Each extension should contribute percepts specific to that agent type as well as common methods (e.g. to perform actions) and properties (such as a position for embodied agents).
The model is instantiated based on information stored within the MASs extension of the OWL2MAS skeleton ontology. Agents in the system correspond to instances of the Agent class and its subclasses, the agent's name within the MAS corresponds exactly to the fragemnt of the individual's URI. The specific class used to model the agent is taken from a special annotation ("AgentClass", which specifies the class's fully-qualified Java name) that is applied to the most specific subclass of Agent that can be inferred for the individual. Additional application-specific properties (such as x,y coordinates for embodied agents) can be specified using sub-properties of the agentModelProperty data property, corresponding code should be added to the agent class constructor to handle such properties.